I-#75: Separated react ui components from logic
