+++ clients/web/.eslintrc.cjs
@@ -23,7 +23,6 @@ module.exports = {
   settings: {
     'import/resolver': {
       alias: {
-        map: [['@', './src']],
         extensions: ['.ts', '.tsx', '.js', '.jsx', '.json'],
       },
     },
