+++ clients/web/src/presentation/pages/index.ts
@@ -4,7 +4,7 @@ export const Pages: Record<
   string,
   React.LazyExoticComponent<React.FC<object>>
 > = {
-  Login: React.lazy(async () => import('./auth/Login')),
-  Signup: React.lazy(async () => import('./auth/Signup')),
+  Login: React.lazy(async () => import('./auth/Login/Login')),
+  Signup: React.lazy(async () => import('./auth/SignUp/Signup')),
   Dashboard: React.lazy(async () => import('./dashboard/Dashboard')),
 };
