+++ clients/web/src/presentation/router/protectedRoutes.tsx
@@ -2,7 +2,7 @@ import { Navigate, Outlet } from 'react-router-dom';
 import useAuthenticatedUser from '@client-commons/hooks/useAuthenticatedUser';
 import FullPageLoader from '../components/loader/FullPageLoader';
 import * as RouteConsts from '../../constants/routeConsts';
-import DefaultLayout from '../layout/DefaultLayout';
+import DefaultLayout from '../layout/DefaultLayout/DefaultLayout';
 
 export function ProtectedRoute() {
   const { user, loading } = useAuthenticatedUser();
